refactor(static analyze): collect all rule and ssa.Option
