+++ .github/workflows/essential-tests.yml
@@ -86,7 +86,7 @@ jobs:
               # SSA
               go test -timeout 20s ./common/yak/yak2ssa ./common/yak/ssaapi 
               # SSA plugin rule/option
-              go test -timeout 20s ./common/yak/plugin_type_analyzer/test 
+              go test -timeout 20s ./common/yak/static_analyzer/test 
               # StaticAnalyze
               go test -timeout 20s -run TestAnalyzeMustPASS* ./common/coreplugin
