+++ common/yak/static_analyzer/plugin_type/plugin_type.go
