+++ common/yak/static_analyzer/rules/risk_rules.go
@@ -1,8 +1,8 @@
 package rules
 
 import (
-	"github.com/yaklang/yaklang/common/yak/plugin_type_analyzer/plugin_type"
 	"github.com/yaklang/yaklang/common/yak/ssaapi"
+	"github.com/yaklang/yaklang/common/yak/static_analyzer/plugin_type"
 )
 
 func init() {
