+++ common/yak/static_analyzer/ssa_option/yak_options.go
@@ -3,9 +3,9 @@ package ssa_option
 import (
 	"reflect"
 
-	"github.com/yaklang/yaklang/common/yak/plugin_type_analyzer/plugin_type"
 	"github.com/yaklang/yaklang/common/yak/ssa"
 	"github.com/yaklang/yaklang/common/yak/ssaapi"
+	"github.com/yaklang/yaklang/common/yak/static_analyzer/plugin_type"
 	"github.com/yaklang/yaklang/common/yak/yaklang"
 )
