+++ common/yak/static_analyzer/test/option_plugin_test.go
