+++ common/yakgrpc/grpc_yaklang_inspect_information.go
@@ -14,9 +14,9 @@ import (
 	"github.com/yaklang/yaklang/common/cve/cveresources"
 	"github.com/yaklang/yaklang/common/log"
 	"github.com/yaklang/yaklang/common/utils"
-	pta "github.com/yaklang/yaklang/common/yak/plugin_type_analyzer"
-	"github.com/yaklang/yaklang/common/yak/plugin_type_analyzer/information.go"
 	"github.com/yaklang/yaklang/common/yak/ssaapi"
+	pta "github.com/yaklang/yaklang/common/yak/static_analyzer"
+	"github.com/yaklang/yaklang/common/yak/static_analyzer/information"
 	"github.com/yaklang/yaklang/common/yakgrpc/yakit"
 	"github.com/yaklang/yaklang/common/yakgrpc/ypb"
 )
